Add copy code button to code samples (#810)
* Refactor sample code presenter into own control

* Add copy sample code button

* Add tooltip to copy sample code button

* Changed text to "Copied!"

* Revert sln version change, fix issue were text was not centered
