+++ XamlControlsGallery/Controls/ControlExample.xaml
@@ -2,12 +2,9 @@
     x:Class="AppUIBasics.ControlExample"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:Core="using:Microsoft.Xaml.Interactions.Core"
-    xmlns:Interactivity="using:Microsoft.Xaml.Interactivity"
-    xmlns:Media="using:Microsoft.Xaml.Interactions.Media"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:contract7Present="http://schemas.microsoft.com/winfx/2006/xaml/presentation?IsApiContractPresent(Windows.Foundation.UniversalApiContract,7)"
+    xmlns:contract7Present="http://schemas.microsoft.com/winfx/2006/xaml/presentation?IsApiContractPresent(Windows.Foundation.UniversalApiContract,7)" xmlns:controls="using:AppUIBasics.Controls"
     x:Name="RootPanel"
     Margin="0,16,0,0"
     d:DesignHeight="250"
@@ -146,29 +143,27 @@
                     Visibility="{x:Bind Options, Converter={StaticResource nullToVisibilityConverter}}" />
             </Grid>
 
-            <!--  XAML display  -->
-            <ScrollViewer
-                x:Name="CodeScrollViewer"
+            <!--  Sample code display  -->
+            <StackPanel x:DefaultBindMode="OneWay"
                 Grid.Row="1"
+                contract7Present:CornerRadius="0,0,8,8"
                 Background="{ThemeResource CardBackgroundFillColorSecondaryBrush}"
                 BorderBrush="{ThemeResource CardStrokeColorDefaultBrush}"
-                BorderThickness="1,0,1,1"
-                HorizontalScrollBarVisibility="Auto"
-                HorizontalScrollMode="Auto"
-                contract7Present:CornerRadius="0,0,8,8"
-                VerticalScrollBarVisibility="Disabled"
-                VerticalScrollMode="Disabled">
-                <StackPanel>
-                    <ContentPresenter
-                        x:Name="XamlPresenter"
-                        Loaded="XamlPresenter_Loaded"
-                        Margin="16,20" />
-                    <ContentPresenter
-                        x:Name="CSharpPresenter"
-                        Loaded="CSharpPresenter_Loaded"
-                        Margin="16,20" />
-                </StackPanel>
-            </ScrollViewer>
+                BorderThickness="1,0,1,1">
+                <controls:SampleCodePresenter
+                    x:Name="XamlPresenter"
+                    Code="{x:Bind Xaml}"
+                    CodeSourceFile="{x:Bind XamlSource}"
+                    Substitutions="{x:Bind Substitutions}"
+                    Margin="16,20" />
+                <controls:SampleCodePresenter
+                    x:Name="CSharpPresenter"
+                    Code="{x:Bind CSharp}"
+                    CodeSourceFile="{x:Bind CSharpSource}"
+                    Substitutions="{x:Bind Substitutions}"
+                    IsCSharpSample="true"
+                    Margin="16,20" />
+            </StackPanel>
 
             <StackPanel x:Name="ScreenshotControls" Grid.Row="2" Orientation="Horizontal" Spacing="6" Margin="0,6,0,0" Visibility="Collapsed">
                 <Button
