+++ XamlControlsGallery/Controls/SampleCodePresenter.xaml
@@ -1,0 +1,41 @@
+<UserControl
+    x:Class="AppUIBasics.Controls.SampleCodePresenter"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" xmlns:controls="using:Microsoft.UI.Xaml.Controls"
+    mc:Ignorable="d"
+    ActualThemeChanged="SampleCodePresenter_ActualThemeChanged"
+    Loaded="SampleCodePresenter_Loaded">
+
+    <Grid>
+        <VisualStateManager.VisualStateGroups>
+            <VisualStateGroup x:Name="ConfirmDialogGroups">
+                <VisualState x:Name="ConfirmationDialogHidden"/>
+                <VisualState x:Name="ConfirmationDialogVisible">
+                    <VisualState.Setters>
+                        <Setter Target="CopyCodeButton.IsEnabled" Value="False"/>
+                        <Setter Target="CopyConfirmationTeachingTip.Visibility" Value="Visible"/>
+                        <Setter Target="CopyConfirmationTeachingTip.IsOpen" Value="True"/>
+                    </VisualState.Setters>
+                </VisualState>
+            </VisualStateGroup>
+        </VisualStateManager.VisualStateGroups>
+        <ScrollViewer
+                x:Name="CodeScrollViewer" Padding="0,4,32,10"
+                HorizontalScrollBarVisibility="Auto"
+                HorizontalScrollMode="Auto"
+                VerticalScrollBarVisibility="Disabled"
+                VerticalScrollMode="Disabled">
+            <ContentPresenter x:Name="CodePresenter" Loaded="CodePresenter_Loaded"/>
+        </ScrollViewer>
+        <Button x:Name="CopyCodeButton" AutomationProperties.Name="Copy sample code"
+                ToolTipService.ToolTip="Copy sample code" VerticalAlignment="Top"
+                HorizontalAlignment="Right" Click="CopyCodeButton_Click" Padding="4">
+            <FontIcon Glyph="&#xE8C8;" FontSize="16"/>
+        </Button>
+
+        <controls:TeachingTip x:Name="CopyConfirmationTeachingTip" x:Load="False" Visibility="Collapsed"
+            Target="{x:Bind CopyCodeButton}" Subtitle="Copied!" IsLightDismissEnabled="True"/>
+    </Grid>
+</UserControl>
